#16 Refactor the App.tsx of the project to separate Monitor function into separate files
