+++ client/src/views/HomeView.tsx
@@ -1,0 +1,9 @@
+import MonitorComponent from "../components/MonitorComponent.tsx"
+
+const HomeView = () => (
+  <div className="home-container">
+    <MonitorComponent />
+  </div>
+);
+
+export default HomeView;
